+++ src/routes/+layout.svelte
@@ -101,8 +101,8 @@
 		updateSettings({ type: 'displayType', value: userSettings.displaySettings.displayType, skipTrackEvent: true });
 	}
 
-	// If wbw language was set to Russian, Ingush or Persian, switch back to English
-	$: if ([9, 10, 19].includes($__wordTranslation)) {
+	// If wbw language was set to Russian or Ingush, switch back to English
+	$: if ([9, 10].includes($__wordTranslation)) {
 		updateSettings({ type: 'wordTranslation', value: 1 });
 	}